requirements.txt
torch
torchvision
numpy
pandas
matplotlib
tqdm
Pillow
scikit-learn
opencv-python
pathlib
argparse
einops
grad-cam
blobfile
mpi4py
imageio
# For the openai logs
tensorflow
